VRSelect, Wiggler: Extract Wiggler and code cleanup
